Added PN noise and signal analysis
Added PN noise in C++ code and added FFT-based signal analysis in Matlab
code.

+++ signalgenerering/pn.m
@@ -1,3 +1,3 @@
 pn_signal = mls(17);
 output_file = fopen('output/mls.bin', 'w')
-fwrite(output_file, pn_signal, 'int')
+fwrite(output_file, pn_signal, 'double')
